Removing all dependant code based on XFS check (#10)
Commit bc6d8e2 removed the check for XFS as well came to a consensus it
could  be removed during migrating to CMake for the build system. With
the check no longer there, there is no reason to keep around dead code
related to XFS anymore.

This removes all code that would have been enabled through the XFS
check. The main thing it did was allow for the -preallocate flag to be
used. However, this was never used for anything other than XFS. If we
decide to revisit this for thing it is work adding back we can update
the build system again to check and add it back in. For now though, it
is just dead code.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>