Only print NUMA nodes if requested
After updating the build system to CMake XDD was outputing the NUMA
domain for each worker thread in the header of the output. This address
that issue by printing that info in the event it is requested through
the `-debug` flag.

Since the `GO_DEBUG_USER1` global debug was never used, it was updated
to be used for init debugging and changed to GO_DEBUG_INIT. So if the
user does `-debug init` they will now see what cpus the worker threads
have been pinned to.

As part of the work, I also reorganized the test directory a bit. I made
a separate debug test directory just to test out this functionality as
it really did not make sense for it to be in the fuctional test cases. I
placed all the common test CMake variables in the CMakeLists.txt and
then just added individual test information in the functional and debug
CMakeLists.txt files.

Fixes #13
Signed-off-by: Brian Atkinson <batkinson@lanl.gov>