Added additional test for -datapattern file/wholefile (#38)
This commit adds another test for -datapattern file/wholefile.
It verifies even when `bs` is smaller than `reqsize`, the code
should still runs without segmentation fault. It also compares
between the input file and output file to ensure it matches and
the offset is correctly calculated with different number of `qd`

Signed-off-by: Cade Gore <79346293+cadegore@users.noreply.github.com>